Extract CalculationContext
